Added Action Overflow Button
- Added Icons to Add Road and Refresh
- Added Overflow Button
- Added Settings Activity

The settings activity will look at the current IP Address where the
device is connected to. It will then give you the possibility to change
the last value in the IP range.
When you click the back button the IP address will be saved in a
SharedPreferences File under "ip_address".

Problem: When a String is set in Strings.xml it can't be changed in
runtime. so it will be better to put the IP Address and maybe the port
in a SharedPreferences file which can be changed and added to during
runtime, a SQLite database is too heavy for this.