Added Longitude and Latitude fields to the View of a single road which is also clickable. If you click it, Maps will now open and put a pin on the map.
The Longitude and Latitude are are still hardcoded Strings. This is because even though I enter the longitude and latitude values in the Constructor of Road.java the values are both null when requested by the ViewSingleRoadFragment.java.
Still Trying to fix this problem.

Also since the server doesn't support Longitude and Latitude values you have to parse them as hardcoded values to the constructor of Road.java